AG-13391 - Move navigator to Enterprise package.
