+++ packages/ag-charts-community/src/main-test.ts
@@ -1,5 +1,7 @@
 // Undocumented APIs used by Enterprise Modules.
 export * from './chart/test/utils';
 export { COMMUNITY_AND_ENTERPRISE_EXAMPLES as GALLERY_EXAMPLES } from './chart/test/examples-gallery';
+export * as CROSSLINE_EXAMPLES from './chart/crossline/test/examples';
 export * from './chart/test/examples';
 export * from './chart/series/test/examples';
+export * from './chart/test/findTarget';
