+++ packages/ag-charts-community/src/module-support.ts
@@ -42,8 +42,6 @@ export * from './module/moduleContext';
 export * from './module/enterpriseModule';
 export * from './chart/background/background';
 export * from './chart/background/backgroundModule';
-export * from './chart/navigator/navigator';
-export * from './chart/navigator/navigatorModule';
 export * from './chart/chartAnimationPhase';
 export * from './chart/chartAxisDirection';
 export * from './chart/axis/axisUtil';
@@ -136,7 +134,7 @@ export { Logger } from './util/logger';
 export { sanitizeHtml } from './util/sanitize';
 export { isContinuous } from './util/value';
 export type { PlacedLabel, PointLabelDatum, LabelPlacement } from './scene/util/labelPlacement';
-export { BBoxValues } from './util/bboxinterface';
+export * from './util/bboxinterface';
 export { Padding } from './util/padding';
 export { lineDistanceSquared } from './util/distance';
 export * from './scale/timeScale';
