+++ packages/ag-charts-community/src/widget/exports.ts
@@ -3,3 +3,4 @@ export * from './widgetEvents';
 export * from './nativeWidget';
 export * from './toolbarWidget';
 export * from './buttonWidget';
+export * from './sliderWidget';
