+++ packages/ag-charts-enterprise/src/axes/__image_snapshots__/time-axis-test-ts-time-axis-examples-when-zooming-for-should-render-as-expected-as-zoom-0-3-0-35-2-snap.png
