+++ packages/ag-charts-enterprise/src/features/navigator/__image_snapshots__/navigator-test-ts-navigator-create-for-nav-zoomed-edge-crosslines-3-it-should-render-to-canvas-as-expected-1-snap.png
