+++ packages/ag-charts-enterprise/src/features/navigator/__image_snapshots__/navigator-test-ts-navigator-create-for-nav-zoomed-no-crosslines-it-should-render-to-canvas-as-expected-1-snap.png
