+++ packages/ag-charts-enterprise/src/features/navigator/navigator.ts
@@ -1,41 +1,212 @@
 import { _ModuleSupport } from 'ag-charts-community';
 
 import { MiniChart } from './miniChart';
+import { type NavigatorButtonType, NavigatorDOMProxy } from './navigatorDOMProxy';
+import { RangeHandle } from './shapes/rangeHandle';
+import { RangeMask } from './shapes/rangeMask';
+import { RangeSelector } from './shapes/rangeSelector';
 
-const { ObserveChanges } = _ModuleSupport;
+const { clamp, BaseModuleInstance, ObserveChanges, Validate, BOOLEAN, POSITIVE_NUMBER } = _ModuleSupport;
 
-export class Navigator extends _ModuleSupport.Navigator {
+export class Navigator extends BaseModuleInstance implements _ModuleSupport.ModuleInstance {
+    // @Validate(OBJECT, { optional: true })
     @ObserveChanges<Navigator, MiniChart>((target, value, oldValue) => {
         target.updateBackground(oldValue?.root, value?.root);
     })
-    override miniChart: MiniChart;
+    public miniChart?: MiniChart;
 
-    public constructor(ctx: _ModuleSupport.ModuleContext) {
-        super(ctx);
+    @Validate(BOOLEAN)
+    @ObserveChanges<Navigator>((target, value) => {
+        target.ctx.zoomManager.setNavigatorEnabled(Boolean(value));
+        target.updateGroupVisibility();
+    })
+    public enabled: boolean = false;
+
+    public mask = new RangeMask();
+    public minHandle = new RangeHandle();
+    public maxHandle = new RangeHandle();
+    private readonly maskVisibleRange = {
+        id: 'navigator-mask-visible-range',
+        getBBox: (): _ModuleSupport.BBox => this.mask.computeVisibleRangeBBox(),
+        toCanvasBBox: (): _ModuleSupport.BBox => this.mask.computeVisibleRangeBBox(),
+        fromCanvasPoint: (x: number, y: number) => ({ x, y }),
+    } satisfies _ModuleSupport.BBoxProvider & { getBBox(): _ModuleSupport.BBox };
+
+    @Validate(POSITIVE_NUMBER)
+    public height: number = 30;
+
+    @Validate(POSITIVE_NUMBER)
+    public spacing: number = 10;
+
+    protected x = 0;
+    protected y = 0;
+    protected width = 0;
+
+    private readonly rangeSelector = new RangeSelector([this.mask, this.minHandle, this.maxHandle]);
+
+    private panStart?: number;
+    private readonly domProxy: NavigatorDOMProxy;
+
+    public constructor(private readonly ctx: _ModuleSupport.ModuleContext) {
+        super();
+
+        this.destroyFns.push(
+            ctx.scene.attachNode(this.rangeSelector),
+            this.ctx.localeManager.addListener('locale-changed', () => this.updateZoom()),
+            this.ctx.layoutManager.registerElement(_ModuleSupport.LayoutElement.Navigator, (e) =>
+                this.onLayoutStart(e)
+            ),
+            this.ctx.layoutManager.addListener('layout:complete', (e) => this.onLayoutComplete(e)),
+            ctx.zoomManager.addListener('zoom-change', (event) => this.onZoomChange(event))
+        );
+
+        this.domProxy = new NavigatorDOMProxy(ctx, this);
+        this.updateGroupVisibility();
 
         this.miniChart = new MiniChart(ctx);
     }
 
-    updateData(data: any) {
-        return this.miniChart.updateData(data);
+    public updateBackground(oldGroup?: _ModuleSupport.Group, newGroup?: _ModuleSupport.Group) {
+        this.rangeSelector?.updateBackground(oldGroup, newGroup);
     }
 
-    processData(dataController: _ModuleSupport.DataController) {
-        return this.miniChart.processData(dataController);
+    private updateGroupVisibility() {
+        const { enabled } = this;
+
+        if (this.rangeSelector == null || enabled === this.rangeSelector.visible) return;
+        this.rangeSelector.visible = enabled;
+        this.domProxy.updateVisibility(enabled);
+
+        if (enabled) {
+            this.updateZoom();
+        } else {
+            this.ctx.zoomManager.updateZoom('navigator');
+        }
     }
 
-    protected override onLayoutStart(opts: _ModuleSupport.LayoutContext) {
-        super.onLayoutStart(opts);
-
+    protected onLayoutStart(ctx: _ModuleSupport.LayoutContext) {
         if (this.enabled) {
+            const { layoutBox } = ctx;
+            const navigatorTotalHeight = this.height + this.spacing;
+            layoutBox.shrink(navigatorTotalHeight, 'bottom');
+            this.y = layoutBox.y + layoutBox.height + this.spacing;
+        } else {
+            this.y = 0;
+        }
+
+        if (this.enabled && this.miniChart) {
             const { top, bottom } = this.miniChart.computeAxisPadding();
-            opts.layoutBox.shrink(top + bottom, 'bottom');
+            ctx.layoutBox.shrink(top + bottom, 'bottom');
             this.y -= bottom;
         }
     }
 
-    override async onLayoutComplete(opts: _ModuleSupport.LayoutCompleteEvent) {
-        await super.onLayoutComplete(opts);
-        await this.miniChart.layout(this.width, this.height);
+    async onLayoutComplete(opts: _ModuleSupport.LayoutCompleteEvent) {
+        const { x, width } = opts.series.rect;
+        const { y, height } = this;
+
+        this.domProxy.updateVisibility(this.enabled);
+        if (this.enabled) {
+            this.layoutNodes(x, y, width, height);
+            this.domProxy.updateBounds({ x, y, width, height });
+        }
+
+        this.x = x;
+        this.width = width;
+
+        await this.miniChart?.layout(this.width, this.height);
+    }
+
+    private canDrag() {
+        const dragStates =
+            _ModuleSupport.InteractionState.Default |
+            _ModuleSupport.InteractionState.Animation |
+            _ModuleSupport.InteractionState.ZoomDrag;
+        return this.enabled && this.ctx.interactionManager.getState() & dragStates;
+    }
+
+    onDragStart(dragging: NavigatorButtonType, { offsetX }: { offsetX: number }) {
+        if (!this.canDrag()) return;
+
+        if (dragging === 'pan') {
+            this.panStart = (offsetX - this.x) / this.width - this.domProxy._min;
+        }
+
+        this.ctx.zoomManager.fireZoomPanStartEvent('navigator');
+    }
+
+    onDrag(dragging: NavigatorButtonType, { offsetX }: { offsetX: number }) {
+        if (!this.canDrag()) return;
+
+        const { panStart, x, width } = this;
+        const { minRange } = this.domProxy;
+        let { _min: min, _max: max } = this.domProxy;
+
+        const ratio = (offsetX - x) / width;
+
+        if (dragging === 'min') {
+            min = clamp(0, ratio, max - minRange);
+        } else if (dragging === 'max') {
+            max = clamp(min + minRange, ratio, 1);
+        } else if (dragging === 'pan' && panStart != null) {
+            const span = max - min;
+            min = clamp(0, ratio - panStart, 1 - span);
+            max = min + span;
+        }
+
+        this.domProxy._min = min;
+        this.domProxy._max = max;
+
+        this.updateZoom();
+    }
+
+    private onZoomChange(event: _ModuleSupport.ZoomChangeEvent) {
+        const { x } = event;
+        if (!x) return;
+
+        this.domProxy.updateMinMax(x.min, x.max);
+        this.updateNodes(x.min, x.max);
+    }
+
+    private layoutNodes(x: number, y: number, width: number, height: number) {
+        const { rangeSelector, mask, minHandle, maxHandle } = this;
+        const { _min: min, _max: max } = this.domProxy;
+
+        rangeSelector.layout(x, y, width, height, minHandle.width / 2, maxHandle.width / 2);
+        mask.layout(x, y, width, height);
+
+        RangeHandle.align(minHandle, maxHandle, x, y, width, height, min, max);
+
+        if (min + (max - min) / 2 < 0.5) {
+            minHandle.zIndex = 3;
+            maxHandle.zIndex = 4;
+        } else {
+            minHandle.zIndex = 4;
+            maxHandle.zIndex = 3;
+        }
+
+        [minHandle, this.maskVisibleRange, maxHandle].forEach((node, index) => {
+            const bbox = node.getBBox();
+            const tbox = { x: bbox.x - x, y: bbox.y - y, height: bbox.height, width: bbox.width };
+            this.domProxy.updateSliderBounds(index, tbox);
+        });
+    }
+
+    private updateNodes(min: number, max: number) {
+        this.mask.update(min, max);
+    }
+
+    private updateZoom() {
+        if (!this.enabled) return;
+        this.domProxy.updateZoom();
+    }
+    updateData(data: any) {
+        return this.miniChart?.updateData(data);
+    }
+
+    async processData(dataController: _ModuleSupport.DataController) {
+        if (this.miniChart) {
+            return this.miniChart?.processData(dataController);
+        }
     }
 }
